pyproject.toml
[project]
name = "airas"
version = "0.0.11.dev10"
description = "Add your description here"
readme = "README.md"
authors = [
    {name="Toma Tanaka", email="ulti4929@gmail.com"}
]
requires-python = ">=3.10,<3.13"
dependencies = [
    "langchain-community>=0.2.15",
    "langchain>=0.2.15",
    "langgraph>=0.2.23",
    "openai>=1.35.13",
    "pydantic>=2.9.2",
    "pypdf>=4.3.1",
    "semanticscholar>=0.8.4",
    "tomli-w>=1.0.0",
    "tomli>=2.0.1",
    "jinja2>=3.1.4",
    "tiktoken>=0.9.0",
    "google-genai>=1.9.0",
    "tenacity>=9.0.0",
    "feedparser>=6.0.11",
    "pytz>=2025.2",
]

[project.urls]
Homepage = "https://www.autores.one/english"
Issues = "https://github.com/auto-res/airas/issues"

[project.optional-dependencies]
mcp = [
    "asyncio>=3.4.3",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/airas"]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "mypy>=1.15.0",
    "nox>=2025.2.9",
    "pre-commit>=3.8.0",
    "pytest>=8.3.3",
    "pytest-cov>=6.1.1",
]

[tool.ruff.lint]
ignore = ["E731","E741","F841","F821"]


[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true
